chore(markov): remove Math.sumPrecise polyfill and unused helpers; merge tests
